#COMMIT AULA 1# (PARCIAL) ASSUNTO: ORDENAÇÃO

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -57,10 +58,14 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
+	public ResponseEntity<?> detalhar(@PathVariable Long id) {
 
-		Topico topico = topicoRepository.getById(id);
-		return new DetalhesDoTopicoDto(topico);
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
 
 	}
 
@@ -68,9 +73,13 @@ public class TopicosController {
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
 
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
 
-		return ResponseEntity.ok(new TopicoDto(topico));
+		return ResponseEntity.notFound().build();
 
 	}
 
@@ -78,8 +87,13 @@ public class TopicosController {
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
 
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+
+		return ResponseEntity.notFound().build();
 
 	}
